Fix documentation in Rubocop cop DocumentedPublicApis
Problem
-------

If you read the docs in the cop, it suggests that you can
put the function documentation string before or after a sorbet
sig.

But if you read the specs, (or try it locally) you get rubocop
errors because it actually requires the comment to come before
the sig.

Solution
---------

Update the docs to match the actual behavior.

The failing spec is `when class defines an instance method with a sig and with documentation below the sig`.
(add GH line link)